Clean Refactor for shorter main
